add PostProcessorTrait
